Move auth to a separate file
